refactor infer_schema.
